+++ server/sonar-web/src/main/js/apps/policy-results/components/App.tsx
