+++ server/sonar-web/src/main/js/apps/policy-results/routes.ts
